backups/CrystalMind_snapshot_20260405_174135/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Materials science
mp-api>=0.41.0
pymatgen>=2024.1.1

# Experiment tracking
mlflow>=2.9.0

# Web demo
streamlit>=1.30.0
plotly>=5.18.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
